🧹 delete ContentEmbedWrapper + move getMp3s() from lib/get to api/contentExtractor
