+++ app/controllers/api/contentExtractor.js
@@ -2,7 +2,37 @@
  * contentExtractor controller (based on contentType controller)
  * Extracts the embedded content in a given http resource
  */
+var path = require('path');
 var get = require('../../lib/get');
+
+var BASE_REG = /<base[^<>]+href\s*=\s*"([^"]*)"[^<>]*>/;
+var MP3_REG = /<a\s+[^<>]*href\s*=\s*"[^"]*\.mp3"/gi;
+var MP3_URL_REG = /<a\s+[^<>]*href\s*=\s*"([^"]*\.mp3)"/i;
+
+function getMp3s() {
+  var mp3s = this.text.match(MP3_REG) || [];
+  var mp3sUniq = [];
+  var base = BASE_REG.test(this.text) ? RegExp.$1 : null;
+  var i, len, mp3;
+  for (i = 0, len = mp3s.length; i < len; i++) {
+    mp3 = mp3s[i].match(MP3_URL_REG)[1];
+    if (mp3.substr(0, 7) !== 'http://') {
+      if (base) mp3 = path.normalize(base + '/' + mp3);
+      else
+        mp3 =
+          'http://' +
+          path.normalize(
+            mp3.charAt(0) === '/'
+              ? this.host + '/' + mp3
+              : this.host + '/' + this.path + '/' + mp3
+          );
+    }
+    mp3s[i] = mp3;
+  }
+  for (i = 0; (mp3 = mp3s[i]); i++)
+    if (mp3sUniq.indexOf(mp3) === -1) mp3sUniq.push(mp3);
+  return mp3sUniq;
+}
 
 exports.controller = function (request, reqParams, response) {
   request.logToConsole('contentExtractor.controller', reqParams);
@@ -14,11 +44,6 @@ exports.controller = function (request, reqParams, response) {
   if (!reqParams || !reqParams.url) {
     console.log('contentExtractor: no url provided => returning null');
     return response.legacyRender(null);
-  }
-
-  function handleError(err) {
-    console.log('contentExtractor error:', err);
-    response.legacyRender({ error: err });
   }
 
   function renderResult(embeds, allEmbeds) {
@@ -44,48 +69,29 @@ exports.controller = function (request, reqParams, response) {
     response.legacyRender({ embeds: embeds, allEmbed: allEmbeds });
   }
 
-  var url = reqParams.url;
-
-  try {
-    get(url, function (err, page) {
-      if (err)
-        //throw err;
-        console.log('contentExtractor.get() error: ', err, err.stack);
-      else
-        page.extractEmbeds(function (embeds) {
-          // 1) extract named embeds
-          var embedRefs = [];
-          embeds = embeds || [];
-          for (let i in embeds)
-            if (embeds[i] && embeds[i].name) embedRefs.push(embeds[i]);
-          console.log(
-            '-> extracted embeds',
-            embedRefs.length,
-            ' (skipped',
-            embeds.length - embedRefs.length,
-            'no-name embeds)'
-          );
-
-          // 2) extract mp3
-          var mp3 = page.getMp3s();
-          if (mp3) {
-            console.log('-> extracted mp3', mp3.length);
-            for (let i in mp3)
-              embedRefs.push({
-                type: 'mp3 file',
-                url: mp3[i],
-                name: mp3[i]
-                  .split('/')
-                  .pop()
-                  .replace(/\.mp3$/i, ''),
-              });
-          }
-          renderResult(embedRefs, embeds);
-        });
-    });
-  } catch (err) {
-    handleError(err);
-  }
+  get(reqParams.url, function (err, page) {
+    if (err) {
+      console.log('contentExtractor.get() error: ', err, err.stack);
+      response.legacyRender({ error: err });
+    } else {
+      const embeds = [];
+      const embedRefs = [];
+      var mp3 = getMp3s.bind(page)();
+      if (mp3) {
+        console.log('-> extracted mp3', mp3.length);
+        for (let i in mp3)
+          embedRefs.push({
+            type: 'mp3 file',
+            url: mp3[i],
+            name: mp3[i]
+              .split('/')
+              .pop()
+              .replace(/\.mp3$/i, ''),
+          });
+      }
+      renderResult(embedRefs, embeds);
+    }
+  });
 };
 
 /*
